frontend/requirements.txt
streamlit
python-dotenv==1.0.0
pymongo
streamlit_authenticator